% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otu_table_duckdb.R
\name{close_otu_table_duckdb}
\alias{close_otu_table_duckdb}
\title{Close DuckDB connection and cleanup}
\usage{
close_otu_table_duckdb(x)
}
\arguments{
\item{x}{An otu_table_duckdb connection}
}
\description{
Close DuckDB connection and cleanup
}
