% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq_duckdb.R
\name{close_phyloseq_duckdb}
\alias{close_phyloseq_duckdb}
\title{Close DuckDB connection and cleanup}
\usage{
close_phyloseq_duckdb(physeq)
}
\arguments{
\item{physeq}{A phyloseq_duckdb object}
}
\description{
Properly close the DuckDB connection and clean up resources. It's important
to call this function when you're done with the phyloseq object to ensure
proper resource management.
}
\examples{
# Create phyloseq object
ps <- phyloseq_duckdb(otu_table = otu_mat)

# Do some analysis
alpha_div <- alpha_diversity_duckdb(ps)

# Clean up when done
close_phyloseq_duckdb(ps)
}
