% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data_duckdb.R
\name{close_sample_data_duckdb}
\alias{close_sample_data_duckdb}
\title{Close DuckDB connection and cleanup}
\usage{
close_sample_data_duckdb(x)
}
\arguments{
\item{x}{A sample_data_duckdb connection}
}
\description{
Close DuckDB connection and cleanup
}
