% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tax_table_duckdb.R
\name{close_tax_table_duckdb}
\alias{close_tax_table_duckdb}
\title{Close DuckDB connection and cleanup}
\usage{
close_tax_table_duckdb(x)
}
\arguments{
\item{x}{A tax_table_duckdb connection}
}
\description{
Close DuckDB connection and cleanup
}
