% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otu_table_duckdb.R
\name{filter_taxa_duckdb}
\alias{filter_taxa_duckdb}
\title{Filter OTU table based on abundance thresholds using DuckDB}
\usage{
filter_taxa_duckdb(x, min_abundance = 1, min_samples = 1)
}
\arguments{
\item{x}{An otu_table_duckdb connection}

\item{min_abundance}{Minimum abundance threshold}

\item{min_samples}{Minimum number of samples where taxa must be present}
}
\value{
Filtered otu_table_duckdb connection
}
\description{
Filter OTU table based on abundance thresholds using DuckDB
}
