% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowsum-methods.R
\name{rowsum,otu_table,ANY,logical-method}
\alias{rowsum,otu_table,ANY,logical-method}
\title{sum values of otu by row, based on a grouping variable. Extended version of the base R function.}
\usage{
\S4method{rowsum}{otu_table,ANY,logical}(x, group, reorder)
}
\arguments{
\item{x}{an otutable}

\item{group}{variable to group by}

\item{reorder}{logical. Whether to reorder taxa by their abundance.}
}
\description{
sum values of otu by row, based on a grouping variable. Extended version of the base R function.
}
