% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otu_table_duckdb.R
\name{sample_sums_duckdb}
\alias{sample_sums_duckdb}
\title{Calculate sample sums for DuckDB OTU table}
\usage{
sample_sums_duckdb(x)
}
\arguments{
\item{x}{An otu_table_duckdb connection}
}
\value{
Named numeric vector of sample sums
}
\description{
Calculate sample sums for DuckDB OTU table
}
