% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tax_glom_duckdb.R
\name{tax_glom_duckdb}
\alias{tax_glom_duckdb}
\title{Agglomerate taxa by taxonomic rank using DuckDB}
\usage{
tax_glom_duckdb(physeq, taxrank)
}
\arguments{
\item{physeq}{A phyloseq_duckdb object}

\item{taxrank}{Character string specifying the taxonomic rank to glom by}
}
\value{
A new phyloseq_duckdb object with agglomerated taxa
}
\description{
This function combines taxa at a specified taxonomic rank, summing their abundances.
Uses DuckDB's efficient aggregation capabilities for memory-efficient processing of
large datasets.
}
\examples{
# Agglomerate at Phylum level
ps_phylum <- tax_glom_duckdb(ps, "Phylum")

# Agglomerate at Class level
ps_class <- tax_glom_duckdb(ps, "Class")
}
\seealso{
\code{\link{filter_taxa_duckdb}}
}
