% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otu_table_duckdb.R
\name{transform_sample_counts_duckdb}
\alias{transform_sample_counts_duckdb}
\title{Transform OTU abundances using DuckDB}
\usage{
transform_sample_counts_duckdb(x, transform = "log10")
}
\arguments{
\item{x}{An otu_table_duckdb connection}

\item{transform}{The transformation to apply ('log10', 'log2', 'sqrt')}
}
\value{
A new otu_table_duckdb connection with transformed values
}
\description{
Apply common transformations to abundance data, such as log or square root.
Operations are performed efficiently using DuckDB's SQL engine.
}
\examples{
# Log transform abundances
otu_log <- transform_sample_counts_duckdb(otu, "log10")
}
\seealso{
\code{\link{filter_taxa_duckdb}}
}
