% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data_duckdb.R
\name{unique_sample_values_duckdb}
\alias{unique_sample_values_duckdb}
\title{Get unique values for a sample metadata variable}
\usage{
unique_sample_values_duckdb(x, variable)
}
\arguments{
\item{x}{A sample_data_duckdb connection}

\item{variable}{Variable to get unique values for}
}
\value{
Vector of unique values
}
\description{
Efficiently retrieve all unique values for a specific sample metadata variable.
Takes advantage of DuckDB's columnar storage for fast distinct value queries.
}
\examples{
# Get all unique treatments
treatments <- unique_sample_values_duckdb(sdata, "Treatment")

# Get all timepoints
times <- unique_sample_values_duckdb(sdata, "Time")
}
\seealso{
\code{\link{filter_sample_data_duckdb}}
}
